tools/assistant/requirements.txt
pyttsx3>=2.90
openai>=0.27.0
pandas>=2.0.0
openpyxl>=3.1.0
pillow>=9.0.0
pytesseract>=0.3.10
tabulate>=0.9.0
twilio>=8.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
sendgrid>=6.9.0
